Stage_Test/MagicLinkProject/app/backend/requirements.txt
flask
gunicorn
firebase-admin
pyjwt
pydantic
requests
